+++ CHANGELOG.md
@@ -7,7 +7,7 @@ the release.
 
 ## Unreleased
 
-* [frontend] reunite trace and add session id to baggage
+* [frontend] reunite trace from loadgenerator
   ([#1506](https://github.com/open-telemetry/opentelemetry-demo/pull/1506))
 * [frauddetectionservice] use span links when consuming from Kafka
   ([#1501](https://github.com/open-telemetry/opentelemetry-demo/pull/1501))